MAJOR UPDATE
Adding the genotyping analysis scripts.
Generalizing the snakemake rules so that the paths can be given in the config file.
Adding initial structure to chimp analysis and vamos analysis.
Matching the structure to the sv-analysis github page so that it is reviewer-friendly and can be copy-pasted.
The snakemake pipelines still need to be tested. They were initially tested only with the dry-run option. They need to be run properly for final testing.